requirements.txt
opencv-python
numpy
scikit-learn
joblib
matplotlib
pandas
tesseract